+++ choco-solver/src/main/java/solver/variables/delta/monitor/SetDeltaMonitor.java
@@ -46,7 +46,7 @@ public class SetDeltaMonitor implements ISetDeltaMonitor {
     protected int[] frozenFirst, frozenLast; // same as previous while the recorder is frozen, to allow "concurrent modifications"
     protected ICause propagator;
 
-    long timestamp = -1;
+    int timestamp = -1;
     final ISearchLoop loop;
 
     public SetDeltaMonitor(ISetDelta delta, ICause propagator) {
